Move the reporting bits into reporting.cpp
