+++ src/app/chip_data_model.gni
@@ -204,6 +204,7 @@ template("chip_data_model") {
 
     if (!chip_build_controller_dynamic_server) {
       sources += [
+        "${_app_root}/reporting/reporting.cpp",
         "${_app_root}/util/DataModelHandler.cpp",
         "${_app_root}/util/attribute-storage.cpp",
         "${_app_root}/util/attribute-table.cpp",
